+++ src/elements/Order.php
@@ -7,8 +7,10 @@
 
 namespace craft\commerce\elements;
 
+use CommerceGuys\Addressing\AddressInterface;
 use Craft;
 use craft\base\Element;
+use craft\base\FieldInterface;
 use craft\commerce\base\AdjusterInterface;
 use craft\commerce\base\Gateway;
 use craft\commerce\base\GatewayInterface;
@@ -1339,7 +1341,7 @@ class Order extends Element
         $datetimeAttributes = array_unique(array_merge($datetimeAttributes, $this->datetimeAttributes()));
 
         foreach ($datetimeAttributes as $attribute) {
-            $fields[$attribute] = static function ($model, $attribute) {
+            $fields[$attribute] = static function($model, $attribute) {
                 if (!empty($model->$attribute)) {
                     $formatter = Craft::$app->getFormatter();
 
@@ -1403,7 +1405,7 @@ class Order extends Element
             // Are the addresses both being set to each other.
             [
                 ['billingAddress', 'shippingAddress'], 'validateAddressReuse',
-                'when' => function ($model) {
+                'when' => function($model) {
                     /** @var Order $model */
                     return !$model->isCompleted;
                 },
@@ -2422,11 +2424,11 @@ class Order extends Element
             $this->_transactions = Plugin::getInstance()->getTransactions()->getAllTransactionsByOrderId($this->id);
         }
 
-        $paidTransactions = ArrayHelper::where($this->_transactions, static function (Transaction $transaction) {
+        $paidTransactions = ArrayHelper::where($this->_transactions, static function(Transaction $transaction) {
             return $transaction->status == TransactionRecord::STATUS_SUCCESS && ($transaction->type == TransactionRecord::TYPE_PURCHASE || $transaction->type == TransactionRecord::TYPE_CAPTURE);
         });
 
-        $refundedTransactions = ArrayHelper::where($this->_transactions, static function (Transaction $transaction) {
+        $refundedTransactions = ArrayHelper::where($this->_transactions, static function(Transaction $transaction) {
             return $transaction->status == TransactionRecord::STATUS_SUCCESS && $transaction->type == TransactionRecord::TYPE_REFUND;
         });
 
@@ -2781,8 +2783,9 @@ class Order extends Element
             throw new InvalidArgumentException('Can not set a shipping address on the order that is is not owned by the order.');
         }
 
-        if ($currentId = $address->id) {
-            if (AddressElement::find()->id($currentId)->ownerId(['not', $this->id])->exists()) {
+        if ($address->id) {
+            $existingAddress = AddressElement::find()->id($address->id)->one();
+            if ($existingAddress->ownerId != $this->id) {
                 throw new InvalidArgumentException('Can not set a shipping address on the order that belongs to another element via ownerId');
             }
         }
@@ -2876,8 +2879,9 @@ class Order extends Element
             throw new InvalidArgumentException('Can not set a billing address on the order that is is not owned by the order.');
         }
 
-        if ($currentId = $address->id) {
-            if (Address::find()->id($currentId)->ownerId(['not', $this->id])->exists()) {
+        if ($address->id) {
+            $existingAddress = AddressElement::find()->id($address->id)->one();
+            if ($existingAddress->ownerId != $this->id) {
                 throw new InvalidArgumentException('Can not set a billing address on the order that belongs to another element via ownerId');
             }
         }
@@ -2896,6 +2900,39 @@ class Order extends Element
     {
         $this->billingAddressId = null;
         $this->_billingAddress = null;
+    }
+
+    /**
+     * Returns whether the billing and shipping addresses' data matches
+     *
+     * @return bool
+     * @since 4.0.4
+     */
+    public function hasMatchingAddresses(): bool
+    {
+        $addressAttributes = (new ReflectionClass(AddressInterface::class))->getMethods();
+        $addressAttributes = array_map(static function(\ReflectionMethod $method) {
+            // Remove `get` and lower case first character
+            return lcfirst(substr($method->name, 3));
+        }, $addressAttributes);
+
+        $customFieldHandles = array_map(static function(FieldInterface $field) {
+            return $field->handle;
+        }, (new AddressElement())->getFieldLayout()->getCustomFields());
+
+        $toArrayHandles = [...$addressAttributes, ...$customFieldHandles];
+
+        $shippingAddress = $this->getShippingAddress();
+        if ($shippingAddress instanceof AddressElement) {
+            $shippingAddress = $shippingAddress->toArray($toArrayHandles);
+        }
+
+        $billingAddress = $this->getBillingAddress();
+        if ($billingAddress instanceof AddressElement) {
+            $billingAddress = $billingAddress->toArray($toArrayHandles);
+        }
+
+        return $billingAddress == $shippingAddress;
     }
 
     /**
@@ -2977,8 +3014,7 @@ class Order extends Element
      */
     public function setPaymentCurrency(
         string $value,
-    ): void
-    {
+    ): void {
         $this->_paymentCurrency = $value;
     }
 
@@ -3008,9 +3044,7 @@ class Order extends Element
     /**
      * Sets the order's selected payment source
      */
-    public function setPaymentSource(
-        ?PaymentSource $paymentSource,
-    ): void
+    public function setPaymentSource(?PaymentSource $paymentSource): void
     {
         if (!$paymentSource instanceof PaymentSource && $paymentSource !== null) {
             throw new InvalidArgumentException('Only a PaymentSource or null are accepted params');
@@ -3035,9 +3069,7 @@ class Order extends Element
     /**
      * Sets the order's selected gateway id.
      */
-    public function setGatewayId(
-        int $gatewayId,
-    ): void
+    public function setGatewayId(int $gatewayId): void
     {
         $this->gatewayId = $gatewayId;
         $this->paymentSourceId = null;
@@ -3335,7 +3367,7 @@ class Order extends Element
             $orphanedAddresses->id($safeIds);
         }
 
-        ($orphanedAddresses->collect())->each(function (AddressElement $address) {
+        ($orphanedAddresses->collect())->each(function(AddressElement $address) {
             Craft::$app->getElements()->deleteElement($address, true);
         });
     }